requirements.txt
alembic==1.14.0
alembic-postgresql-enum==1.4.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.29.0
black==24.10.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
databases==0.7.0
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
identify==2.6.2
idna==3.10
isort==5.13.2
Mako==1.3.6
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nodeenv==1.9.1
ormar==0.20.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
psycopg2-binary==2.9.10
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.41.3
tomli==2.1.0
typing_extensions==4.12.2
uvicorn==0.32.1
virtualenv==20.27.1
aiosmtplib==3.0.2